Change opensearch-tar-install.sh to use 'exec' rather than 'bash' to start up the OpenSearch process
Currently, the opensearch-tar-install.sh script does not correctly pass SIGTERM and SIGKILL down to the child process. Since that line is the last one in this script, the easiest solution is to use 'exec' so that the current process gets replaced by the OpenSearch Java process.

This commit also removes the 'chmod /dev/shm' because that has to be executed as root. This script should not be run as root.

Signed-off-by: Cameron Skinner <camerski@amazon.com>