[Build] bundle OpenSearch Dashboards
Produce an OpenSearch Dashboards bundle with plugins provided by a manifest.

This requires the manifest to be in order with plugins first and then Dashboards
being last. The plugins will need to pulled down and placed in the plugins
folder. Then we need to bootstrap to generate the modules so that we can actually
install the plugins to the release build.

What is missing or not verified
* Environment setup, this requires nvm and yarn to be already set up.
We will need to set that up.
* Dropping the custom config into the bundle
* Signing of the artifact
* Tests

Related Issue:
https://github.com/opensearch-project/opensearch-build/issues/158

Signed-off-by: Kawika Avilla <kavilla414@gmail.com>