Parallel bundle build notifications
With the move to parallel builds, without specificing an agent Jenkins was unable to send notifications.
Reusing the node from the x86 for both success/failure flows resolved this, tested on a fake job extensively.

Adjusted the way that jobs identified themselves, reusing the same stash caused the data to be overwritten, by
using job name to seperate the source this avoided any complex concurrency limitations.  In the future this
can be further improved by generating the architecture builds from script so the names do not need to be
duplicated.

Signed-off-by: Peter Nied <petern@amazon.com>