[Build] add arch arg to build workflow
Build target accepts an arch already so this enables passing the
arch while running the build workflow. This allows building target
operating systems that are different from the operating system of
the build agent.

Use case: chromium package not available on ARM64 but it is just
used for building so if we can just build on x64 then assemble
on ARM64 we can get a bundle that we are expecting.

Updated Jenkinsfile for OpenSearch Dashboards to utilize this
functionality.

Issue resolved:
https://github.com/opensearch-project/opensearch-build/issues/739

Signed-off-by: Kawika Avilla <kavilla414@gmail.com>