[Bundle] post-install for min
Adding install script for OpenSearch and OpenSearch Dashboards so
the bundles include default files.

These files include:
* opensearch-tar-install.sh
* config/opensearch_dashboards.yml

However, we did not want to include opensearch-tar-install.sh into
the OpenSearch Dashboards bundle because the script will fail to
execute as it is only expected to be an OpenSearch project.

Issue resolved:
https://github.com/opensearch-project/opensearch-build/issues/610

Signed-off-by: Kawika Avilla <kavilla414@gmail.com>