[Build] support building targets for platform and arch inputs
Adding `platform` and `arch` as a build argument to be passed to
the build target. This enables cross-platform cross-arch builds,
for example, we can build linux x64 on a linux ARM64 machine.

We would want to do this because some of the libraries that are
required for building (not for runtime) are not available for
certain operating systems.

Also updated the OpenSearch Dashboards Jenkinsfile to pass the
platform and arch.

Issue resolved:
https://github.com/opensearch-project/opensearch-build/issues/739

Signed-off-by: Kawika Avilla <kavilla414@gmail.com>