Seperate check for build from distribution build workflow
Create a new job that will only check if the SHA for the current build file has been created, if so it will trigger the distribution build job.  This should allow for build issues to be discovered more quickly and reduce the amount of no-op runs seen in the distribution-build job in jenkins.

Note; this does rename the job files and so Jenkins updates will need to be make when this is merged.

Signed-off-by: Peter Nied <petern@amazon.com>