Add sha512 generation to promotion job
The promotion workflow is picking up artifacts from a distribution build and placing them in their final resting place on the artifacts server.  This is a perfect oppertunity to do any final polishing steps on artifacts that are ready for release.  Adding this hook for file actions and creating a sha512 generation closure.

This is untested on jenkins, but is the core POC needed to bootstrap the signing workflow.

Signed-off-by: Peter Nied <petern@amazon.com>