Change the issue title for release to be 'proposed'
The current langauge makes the autocut release tickets seem like a
commitment for a release, whereas they are a best guess effort
that a release could happen.  E.g. there might ne a patch release in
flight and then don't release that in favor of a minor release that
also includes the contents as the patch release.

Signed-off-by: Peter Nied <petern@amazon.com>