[Dashboards] integ-test in Jenkins
Execute the integ-test workflow for OpenSearch Dashboards including
tests.

As opposed to OpenSearch, we need an image with the appropriate libraries
so we pass the image but it does not have the wget so the steps were broken
to access the build manifest.

Hardcoding the OpenSearch build id for now until this get resolves:
https://github.com/opensearch-project/opensearch-build/issues/1492

ARM tests will fail still until this is resolved:
https://github.com/opensearch-project/opensearch-build/issues/1381

Issue partially resolved:
https://github.com/opensearch-project/opensearch-build/issues/704

Signed-off-by: Kawika Avilla <kavilla414@gmail.com>