[Tests][Integ] Fix without-security logic if no security plugin
When the integ test workflow runs it needs to start a cluster for
OpenSearch tests, it needs to start a cluster and an
OpenSearch Dashboards for OpenSearch Dashboards tests.

During this workflow, even if we pass `without-security` to the
test manifest for `integ-test` config the workflow makes the
assumption that the security plugin is actually there and tries
to uninstall it or add configs to a plugin that isn't there. Which
fails to start a cluster or a cluster and OpenSearch Dashboards

Issue resolved:
https://github.com/opensearch-project/opensearch-build/issues/1814

Signed-off-by: Kawika Avilla <kavilla414@gmail.com>