[BUILD][OSD] Support version qualifier
Support version qualifier builds for OpenSearch Dashboards and
OpenSearch Dashboards Plugins.

This will create a zip for plugins with the OpenSearch Dashboards
version + qualifier. But the version of the plugin will remain the
same whilst opensearchDashboardsVersion in the opensearch_dashboards.json
will be the OpenSearch Dashboards version, eg, 1.3.0-alpha1. This is why
the version check needed to be modified a little to ensure that it can
get the version + qualifier and the default version of the plugin.

Updated for plugins as well.

Was able to successfully build and connect OSD 1.3.0-alpha1 to
the release version of OS 1.3.0.

Issue resolved:
https://github.com/opensearch-project/OpenSearch-Dashboards/issues/1340

Issue partially resolved:
https://github.com/opensearch-project/opensearch-build/issues/1632

Signed-off-by: Kawika Avilla <kavilla414@gmail.com>