remove OpenSearch `build.sh`: moved to OpenSearch repo
`build.sh` has been copied over to all active branches(*) of the
OpenSearch repo and can now consequently be deleted here. this means,
that from now on, the scripts from the respository will be picked up.

for more information on the order of evaluation of the custom build
scripts, see the [documentation][].

this gives the advantage to have branch-specific new logic in place
there. this has already been used to add the no-jdk targets to the
builds on main and 2.x.

two tests were based on the assumption that there would be a `build.sh`
for OpenSearch in this repository, however these are generic tests and
just used OpenSearch as an example. they have been switched to be based
on OpenSearch-Dashboards instead which - for the time being - still has
its own `build.sh` here.

(*): the following branches have been considered as active:
- main
- 2.x
- 2.3
- 1.x
- 1.3

this is part of opensearch-project/opensearch-build#99

[documentation]: https://github.com/opensearch-project/opensearch-build/tree/main/src/build_workflow#custom-build-scripts

Signed-off-by: Ralph Ursprung <Ralph.Ursprung@avaloq.com>