Restrict access to OpenSearch configuration files
The configuartion files installed by the opensearch package can contain
sensitive information such as clear-text passwords (e.g. to reach an
LDAP directory to authenticate users).  When installing from a tarball,
the configuration file are only accessible to the owner and members of
the group of the file.

Adjust the post-installation script to setup permissions for these files
similar to the ones of the tarball.

Currently, OpenSearch create files in the configuartion directory on
first start (opensearch.keystore), so we cannot change the configuration
directory ownership to prevent the service from tinkering with these
files.

Signed-off-by: Romain Tartière <romain@blogreen.org>