Adjust ownership and permissions of the log directory
The OpenSearch logs may contain sensitive data.  In order to prevent
random users from accessing log files, Debian packages often setup the
log directory so that only the service can write to it, members of the
adm group can read these files, and other users are not granted access.

Adjust the log directory owner and permissions to match that behavior.

Signed-off-by: Romain Tartière <romain@blogreen.org>