Specify that this does not impact 1.x
This will not be backported to 1.x packages, but since 1.x and 2.x
currently share the tooling, add information so that users are not
confused.

As users of 1.x are expected to switch to a newer version at some point,
include the informational message for them to prepare the sooner.

Signed-off-by: Romain Tartière <romain@blogreen.org>